move plotting to separate thread.
When opening a menu the system is delayed for the amount of time
the menu is open. When plotting the delay is played back, but when
plotting is disabled the video signal freezes until it has caught
up. Not sure why or how. Just disable plotting once the orientation
is verified