fix duplicate IMU data appearing.
The framegrabber sometimes returned consecutive frames
instantly, which resulted in IMU values appearing twice.

IMU data is now stored in variable while a test is conducted,
with a time stamp that is compared to the frame time stamp. This
may result in a slight mismatch between IMU data and frame time
but not much.